Remove bad fix
I don't think this was fixing anything but actually breaking things:

* This is just changing docker image defaults to a different folder, and
it's actually doing incorrectly because it's leaving `GEM_HOME` with the
default value, so when installing gems via `gem install <gem>`, they get
installed to a wrong location (this affects generator specs in decidim).
See
https://github.com/docker-library/ruby/blob/master/Dockerfile-debian.template#L69-L74.

* Also, the folder name is confusing, since gems get actually installed
to a "gems" subfolder inside this "gems" folder.

* Finally, I don't see how the linked issue relates to this.